Reorganized directory structure, to eliminate subpackages
