+++ benchmarking/backendSpeed/test_backend_speed.py
@@ -10,10 +10,10 @@ import torch
 import time
 import pickle
 
-from sns_toolbox.design.neurons import SpikingNeuron
-from sns_toolbox.design.connections import SpikingSynapse
-from sns_toolbox.design.networks import Network
-from sns_toolbox.simulate.backends import SNS_Numpy, SNS_Torch, SNS_Sparse, SNS_Manual
+from sns_toolbox.neurons import SpikingNeuron
+from sns_toolbox.connections import SpikingSynapse
+from sns_toolbox.networks import Network
+from sns_toolbox.backends import SNS_Numpy, SNS_Torch, SNS_Sparse, SNS_Manual
 
 """
 ########################################################################################################################
