+++ example_figures/figure_example_net.py
@@ -6,11 +6,11 @@ January 27th 2022
 import numpy as np
 import matplotlib.pyplot as plt
 
-from sns_toolbox.design.neurons import NonSpikingNeuron
-from sns_toolbox.design.connections import NonSpikingSynapse
-from sns_toolbox.design.networks import Network
+from sns_toolbox.neurons import NonSpikingNeuron
+from sns_toolbox.connections import NonSpikingSynapse
+from sns_toolbox.networks import Network
 
-from sns_toolbox.simulate.backends import SNS_Numpy
+from sns_toolbox.backends import SNS_Numpy
 
 spiking = True
 delay = True
