+++ sns_toolbox/compilers.py
@@ -1,5 +1,5 @@
-from sns_toolbox.simulate.backends import SNS_Numpy, SNS_Torch, SNS_Sparse, SNS_Manual
-from sns_toolbox.design.neurons import SpikingNeuron, NonSpikingNeuronWithGatedChannels
+from sns_toolbox.backends import SNS_Numpy, SNS_Torch, SNS_Sparse, SNS_Manual
+from sns_toolbox.neurons import SpikingNeuron, NonSpikingNeuronWithGatedChannels
 
 import numpy as np
 import torch
