+++ sns_toolbox/neurons.py
@@ -11,7 +11,7 @@ from typing import Dict, Any
 import warnings
 import numbers
 
-from sns_toolbox.design.design_utilities import valid_color, set_text_color
+from sns_toolbox.utilities import valid_color
 
 import numpy as np
 import torch
